+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - ✨(backend) capture a posthog event on template based creation
 - ✨(backend) capture posthog events on item deletion
+- ✨(backend) capture a posthog event on conversion request
 
 ### Fixed